recipes/custom_ai_assistant/requirements.txt
openvino==2024.2
optimum-intel==1.16.1
nncf==2.11.0

onnx==1.16.0
onnxruntime==1.16.0

transformers==4.40.2
librosa==0.10.2

gradio==4.31.3